+++ snooty/test_util.py
@@ -60,6 +60,7 @@ def test_get_files() -> None:
         Path("test_data/test_devhub/snooty.toml"),
         Path("test_data/test_intersphinx/snooty.toml"),
         Path("test_data/test_landing/snooty.toml"),
+        Path("test_data/test_parser_failure/snooty.toml"),
     }