+++ snooty/parser.py
@@ -236,7 +236,7 @@ class JSONVisitor:
         elif isinstance(node, docutils.nodes.bullet_list):
             self.state.append(n.ListNode((line,), [], n.ListEnumType.unordered))
         elif isinstance(node, docutils.nodes.enumerated_list):
-            self.state.append(n.ListNode((line,), [], node["enumtype"]))
+            self.state.append(n.ListNode((line,), [], n.ListEnumType[node["enumtype"]]))
         elif isinstance(node, docutils.nodes.list_item):
             self.state.append(n.ListNodeItem((line,), []))
         elif isinstance(node, docutils.nodes.title):