+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve string-handling in spec-parser (DOP-1148).
 
+### Added
+
+- fail_on_diagnostics toml flag (DOP-1132)
+
 ## [v0.4.11] - 2020-06-22
 
 ### Fixed