+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - fail_on_diagnostics toml flag (DOP-1132)
+- Support additional ordered list labels
 
 ## [v0.4.11] - 2020-06-22