+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fail_on_diagnostics toml flag (DOP-1132)
 - Support additional ordered list labels (DOP-1129)
+- Support ordered list "start" attribute (DOP-1210)
 
 ## [v0.4.11] - 2020-06-22