+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support additional ordered list labels (DOP-1129)
 - Support ordered list "start" attribute (DOP-1210)
 
+### Changed
+
+- Improve string-handling in spec-parser (DOP-1148).
+
 ## [v0.4.11] - 2020-06-22
 
 ### Fixed